lc3-vm: add case for op_lea and handle reserved ops differently
I had missed adding op_lea, as a result of which the program would
default to abortion.

It seems that in the files I am testing with (rogue and 2048), the
operations RTI and RES make at least 1 occurrence. As a result, instead
of ending itself, the program should just ignore these instructions. So
do that.